requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0 